Add Vtable
